+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.5 ()
 
 * Set default image size.
+* Add custom post type to ACF post type select.
 
 ## 0.5.4 (2023-06-26)